+++ .gitignore
@@ -2,5 +2,3 @@ node_modules
 # Keep environment variables out of version control
 
 prisma/migrations
-
-assets
