+++ src/module/admin/admin.controller.js
@@ -1,4 +1,5 @@
 const express = require("express");
+const { multer } = require("../../db/index");
 const validator = require("validator");
 const {
   allUsers,
@@ -6,66 +7,83 @@ const {
   deleteUserById,
   createUser,
   editUserById,
+  userUpdate,
 } = require("./admin.service");
 const router = express.Router();
+const images = multer.diskStorage({
+  destination: (req, file, cb) => {
+    cb(null, "assets/admin");
+  },
+  filename: (req, file, cb) => {
+    cb(null, new Date().getTime() + "-" + file.originalname);
+  },
+});
+
+const uploading = multer({ storage: images });
 
 router.get("/", async (req, res) => {
-  const users = await allUsers();
-
-  res.send(users);
-});
-router.post("/", async (req, res) => {
   try {
-    const newData = req.body;
-    const strongPassword = validator.isStrongPassword(newData.passwordUser);
-    if (!strongPassword) {
-      return res.status(400).send({ message: "password not strong" });
-    }
-    if (!req.file) {
-      return res.status(422).send({ message: "image must be uploaded" });
-    }
-    const image = req.file.path;
-    const user = await createUser(newData, image);
+    const users = await allUsers();
 
-    res.send({
-      data: user,
-      message: "create Users success",
-    });
-  } catch (err) {
-    res.status(400).send(err.message);
+    res.status(200).send(users);
+  } catch (error) {
+    return res.status(500).send({ message: error.message });
   }
 });
-router.put("/:id", async (req, res) => {
-  const userId = req.params.id;
-  const newData = req.body;
-  if (
-    !(
-      newData.nameUser &&
-      newData.emailUser &&
-      newData.passwordUser &&
-      newData.tlpUser &&
-      newData.addressUser &&
-      newData.levelUser &&
-      newData.statusUser
-    )
-  ) {
-    return res.status(400).send("some fields are missings");
-  }
-  const strongPass = validator.isStrongPassword(newData.passwordUser);
-  if (!strongPass) {
-    return res.status(400).send({ message: "password not strong" });
-  }
-  if (!req.file) {
-    return res.status(422).send({ message: "image must be uploaded" });
-  }
-  const image = req.file.path;
-  const user = await editUserById(parseInt(userId), newData, image);
+router.post("/add", uploading.single("fotoUser"), createUser);
+// router.post("/", async (req, res) => {
+//   try {
+//     const newData = req.body;
+//     const strongPassword = validator.isStrongPassword(newData.passwordUser);
+//     if (!strongPassword) {
+//       return res.status(400).send({ message: "password not strong" });
+//     }
+//     if (!req.file) {
+//       return res.status(422).send({ message: "image must be uploaded" });
+//     }
+//     const image = req.file.path;
+//     const user = await createUser(newData, image);
 
-  res.send({
-    data: user,
-    message: "Update User success",
-  });
-});
+//     res.send({
+//       data: user,
+//       message: "create Users success",
+//     });
+//   } catch (err) {
+//     res.status(400).send(err.message);
+//   }
+// });
+router.put("/update/:id", uploading.single("fotoUser"), userUpdate);
+// router.put("/:id", async (req, res) => {
+//   const userId = req.params.id;
+//   const newData = req.body;
+//   if (
+//     !(
+//       newData.nameUser &&
+//       newData.emailUser &&
+//       newData.passwordUser &&
+//       newData.tlpUser &&
+//       newData.addressUser &&
+//       newData.levelUser &&
+//       newData.statusUser
+//     )
+//   ) {
+//     return res.status(400).send("some fields are missings");
+//   }
+//   const strongPass = validator.isStrongPassword(newData.passwordUser);
+//   if (!strongPass) {
+//     return res.status(400).send({ message: "password not strong" });
+//   }
+//   if (!req.file) {
+//     return res.status(422).send({ message: "image must be uploaded" });
+//   }
+//   const image = req.file.path;
+//   const user = await editUserById(parseInt(userId), newData, image);
+
+//   res.send({
+//     data: user,
+//     message: "Update User success",
+//   });
+// });
 router.patch("/:id", async (req, res) => {
   try {
     const userId = req.params.id;
