+++ src/module/admin/admin.service.js
@@ -6,7 +6,9 @@ const {
   updateUser,
   insertUser,
 } = require("./admin.repository");
+const validator = require("validator");
 const fs = require("fs");
+
 const allUsers = async () => {
   const users = await findAllUser();
   return users;
@@ -19,8 +21,63 @@ const userById = async (id) => {
   }
   return users;
 };
+const createUser = async (req, res) => {
+  try {
+    const newData = req.body;
+    const strongPassword = validator.isStrongPassword(newData.passwordUser);
+    if (!strongPassword) {
+      return res.status(400).send({ message: "password not strong" });
+    }
+    if (!req.file) {
+      return res.status(422).send({ message: "image must be uploaded" });
+    }
+    const image = req.file.path;
+    const user = await checkCreateUser(newData, image);
 
-const createUser = async (newData, image) => {
+    res.send({
+      data: user,
+      message: "create Users success",
+    });
+  } catch (err) {
+    res.status(400).send(err.message);
+  }
+};
+const userUpdate = async (req, res) => {
+  try {
+    const userId = req.params.id;
+    const newData = req.body;
+    if (
+      !(
+        newData.nameUser &&
+        newData.emailUser &&
+        newData.passwordUser &&
+        newData.tlpUser &&
+        newData.addressUser &&
+        newData.levelUser &&
+        newData.statusUser
+      )
+    ) {
+      return res.status(400).send("some fields are missings");
+    }
+    const strongPassword = validator.isStrongPassword(newData.passwordUser);
+    if (!strongPassword) {
+      return res.status(400).send({ message: "password not strong" });
+    }
+    if (!req.file) {
+      return res.status(422).send({ message: "image must be uploaded" });
+    }
+    const image = req.file.path;
+    const user = await editUserById(parseInt(userId), newData, image);
+
+    res.send({
+      data: user,
+      message: "Update User success",
+    });
+  } catch (err) {
+    res.status(400).send(err.message);
+  }
+};
+const checkCreateUser = async (newData, image) => {
   const check = await checkUser(newData);
   if (check) {
     throw Error("User already exists");
@@ -68,4 +125,5 @@ module.exports = {
   deleteUserById,
   createUser,
   editUserById,
+  userUpdate,
 };
